Move ToCharArray methods to ObsoleteAPI for removal in 4.8.0 RC, #1024
